fix: code reorganization
