+++ packages/react/src/transformationParsers/parseQuality.ts
@@ -1,3 +1,3 @@
-import { Quality } from '../types';
+import { Quality } from '../transformationTypes/quality';
 
 export const parseQuality = (quality: Quality): `q_${string}` => `q_${quality}`;
