+++ packages/react/src/transformationParsers/parseWidth.ts
@@ -1,5 +1,3 @@
-import { WidthOption } from '../types';
+import { WidthOption } from '../transformationTypes/resize';
 
 export const parseWidth = (width: WidthOption) => `w_${width}`;
-
-
