+++ packages/react/src/transformationTypes/aspectRatio.ts
@@ -1,0 +1,3 @@
+export interface AspectRatio {
+  aspectRatio: `${number}:${number}` | number;
+}
