+++ packages/react/src/transformationTypes/effect.ts
@@ -1,0 +1,11 @@
+type SepiaEffect = {
+  type: 'sepia';
+  level?: number;
+};
+
+type BackgroundRemovalEffect = {
+  type: 'backgroundRemoval';
+  mode?: 'fineEdges';
+};
+
+export type Effect = SepiaEffect | BackgroundRemovalEffect;
