+++ packages/react/src/transformationTypes/format.ts
@@ -1,0 +1,20 @@
+export type ImageFormat =
+  | 'auto'
+  | 'jpg'
+  | 'png'
+  | 'gif'
+  | 'webp'
+  | 'bmp'
+  | 'ico'
+  | 'pdf'
+  | 'tiff'
+  | 'eps'
+  | 'jpc'
+  | 'jp2'
+  | 'psd'
+  | 'svg'
+  | 'avif'
+  | 'heic'
+  | 'heif';
+
+export type VideoFormat = 'webm' | 'mp4' | 'mkv' | 'flv' | 'mov' | '3gp' | 'avi' | 'wmv';
