+++ packages/react/src/transformationTypes/gravity.ts
@@ -1,0 +1,26 @@
+// FIXME add custom gravity support
+export type Gravity =
+  | 'north'
+  | 'north_west'
+  | 'north_east'
+  | 'south'
+  | 'south_west'
+  | 'south_east'
+  | 'west'
+  | 'east'
+  | 'center'
+  | 'face'
+  | 'faces'
+  | 'auto'
+  | 'adv_face'
+  | 'adv_faces'
+  | 'custom'
+  | 'custom:face'
+  | 'custom:faces'
+  | 'custom:auto'
+  | 'custom:adv_face'
+  | 'custom:adv_faces'
+  | 'ocr_text'
+  | 'body'
+  | 'liquid'
+  | 'auto:subject';
